requirements.txt
torch==1.8.0
transformers==4.7.0
pyyaml
py-rouge
nltk